# feat(perf): add encoder scenario