--- a/benchmarks/encoder/benchmark
+++ b/benchmarks/encoder/benchmark
@@ -14,11 +14,13 @@ RESULTS_2=/artifacts/${RUN_ID}/${DDTRACE_GIT_COMMIT_ID_2}.json
 PYTHONPATH=/app/.venv_1/lib/python3.9/site-packages \
     python scenario.py \
     --copy-env \
+    --fast \
     -o ${RESULTS_1}
 
 PYTHONPATH=/app/.venv_2/lib/python3.9/site-packages \
     python scenario.py \
     --copy-env \
+    --fast \
     -o ${RESULTS_2}
 
 pyperf compare_to --table ${RESULTS_2} ${RESULTS_1}