--- a/benchmarks/encoder/entrypoint
+++ b/benchmarks/encoder/entrypoint
@@ -1,0 +1,18 @@
+#!/bin/bash
+
+set -o errexit
+set -o pipefail
+set -o nounset
+
+python3 -m venv /app/.venv_1
+python3 -m venv /app/.venv_2
+
+source /app/.venv_1/bin/activate
+pip install git+https://github.com/Datadog/dd-trace-py@${DDTRACE_GIT_COMMIT_ID_1}
+deactivate
+
+source /app/.venv_2/bin/activate
+pip install git+https://github.com/Datadog/dd-trace-py@${DDTRACE_GIT_COMMIT_ID_2}
+deactivate
+
+exec "$@"