--- a/benchmarks/encoder/scenario.py
+++ b/benchmarks/encoder/scenario.py
@@ -7,6 +7,8 @@ from ddtrace.internal.encoding import Encoder
 
 
 try:
+    # the introduction of the buffered encoder changed the internal api
+    # see https://github.com/DataDog/dd-trace-py/pull/2422
     from ddtrace.internal._encoding import BufferedEncoder
 
     def init_encoder(max_size=8 << 20, max_item_size=8 << 20):