## Description

This adds a component performance testing scenario for the encoder. It compares two versions of the library on variations of trace data to be encoded.

To run the scenario:

```sh
# change directory to the benchmarks folder
$ docker build -t perf-encoder --build-arg BENCHMARK=encoder -f base.Dockerfile .
$ docker run -it -e DDTRACE_GIT_COMMIT_ID_1=v0.50.0 -e DDTRACE_GIT_COMMIT_ID_2=v0.49.4 perf-encoder
```

The results from this scenario compare only function timings. System metrics are not collected and compared as of yet. The focus here has been on getting stable timings, using [pyperf](https://pyperf.readthedocs.io/).

The output at the end of the run will compare the two versions. In this case, I'm comparing two versions in which the encoder had not changed. As we expect, the comparison shows no difference.

| Benchmark                                  | v0.49.4 | v0.50.0               |
|--------------------------------------------|---------|-----------------------|
| ntraces:1\|nspans:1000                     | 547 us  | 483 us: 1.13x faster  |
| ntraces:100\|nspans:1000                   | 55.8 ms | 51.6 ms: 1.08x faster |
| ntraces:100\|nspans:100                    | 5.66 ms | 5.05 ms: 1.12x faster |
| ntraces:100\|nspans:100\|ntags:1\|ltags:16 | 9.87 ms | 8.49 ms: 1.16x faster |
| ntraces:100\|nspans:100\|nmetrics:1        | 8.04 ms | 10.3 ms: 1.28x slower |
| ntraces:100\|nspans:100\|nmetrics:100      | 70.8 ms | 94.7 ms: 1.34x slower |
| Geometric mean                             | (ref)   | 1.01x slower          |

## Checklist
- [ ] Added to the correct milestone.
- [ ] Tests provided or description of manual testing performed is included in the code or PR.
- [ ] Library documentation is updated.
- [ ] [Corp site](https://github.com/DataDog/documentation/) documentation is updated (link to the PR).